requirements.txt
# 핵심 패키지
langchain-community==0.3.23
langchain-core==0.3.56
langchain-huggingface>=0.0.6
langchain-chroma>=0.0.10
sentence-transformers>=2.2.2

google-generativeai==0.8.5

# 추가
google-genai==1.15.0 

# PyTorch (GPU 버전)
# cpu는 밑에 줄 주석처리
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

# 벡터 데이터베이스
chromadb>=0.4.18

# 설정 관리
pydantic_settings

# 임베딩 처리에 필요한 최소 유틸리티
numpy>=1.24.0

# PyTorch with GPU support (for CUDA 11.8)
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

##구조화된 출력
google-genai